+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   to Autosubmit repository, merging the projects again #2052
 - Documentation about `FOR.NAME` with values that are not strings #2515
 - Improvement of the error messages for YAML files #2488 and for RO-CRATE #2572
+- Added `--force` flag to `autosubmit pklfix` command, and `--yes` flag to `autosubmit stop` to
+  automatically answer yes to prompts #2569
 
 ### 4.1.15: Bug fixes, enhancements, and new features